requirements.txt
python-dotenv
requests
./pycord
pandas
pytest